% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrono_trigger.R
\name{chrono_trigger}
\alias{chrono_trigger}
\title{Play a short sound from Chrono Trigger}
\usage{
chrono_trigger(sound = 1, expr = NULL)
}
\arguments{
\item{sound}{character string or number specifying what sound to be played
by either specifying one of the built in sounds, specifying the path to a
wav file or specifying an url. The default is 1. Possible sounds are:
\enumerate{ \item \code{"save"} \item \code{"draw"} \item
\code{"leenes_bell"} \item \code{"croak"} \item \code{"chest"}
\item \code{"epoch"} \item \code{"strike"} \item \code{"temporal"}
\item \code{"timegate"} \item \code{"wrong"} \item
\code{"black_tyrano"} \item \code{"continuum"} \item \code{"destruction"}
\item \code{"water_drip"} \item \code{"grunt"} \item \code{"gulls"} \item
\code{"lavos"} \item \code{"robo"} }
If \code{sound} does not match any of the sounds above, or is a valid path
or url, a random sound will be played. Currently \code{chrono_trigger} can
only handle http urls, https is not supported.}

\item{expr}{An optional expression to be excecuted before the sound.}
}
\description{
\code{chrono_trigger} plays a short sound which is useful if you want to get
  notified, for example, when a script has finished. As an added bonus there
  are a number of different sounds to choose from.
If \code{chrono_trigger} is not able to play the sound a warning is issued
  rather than an error. This is in order to not risk aborting or stopping the
  process that you wanted to get notified about.
}
\examples{
# Play a "save" sound
chrono_trigger()

\dontrun{
# Play "gulls" instead of a "save".
chrono_trigger("gulls")
# or
chrono_trigger(16)

# Play a random sound
chrono_trigger(0)

# Update all packages and "ping" when it's ready
update.packages(ask=FALSE); chrono_trigger()
}
}
