% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_wav_fname.R
\name{is_wav_fname}
\alias{is_wav_fname}
\title{Determine if file is a .wav}
\usage{
is_wav_fname(fname)
}
\arguments{
\item{fname}{filename of a .wav sound file}
}
\description{
Checks to see if the filename supplied is a .wav file
}
