% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_audio.R
\name{play_audio}
\alias{play_audio}
\title{Play wave sound}
\usage{
play_audio(fname)
}
\arguments{
\item{fname}{filename of a .wav sound file}
}
\description{
Uses Audio to play a sound effect if available
}
