% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_file.R
\name{play_file}
\alias{play_file}
\title{Controller to play sound}
\usage{
play_file(fname)
}
\arguments{
\item{fname}{filename of a .wav sound file}
}
\description{
Uses system appropriate player if available
}
