% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warcraft.R
\name{warcraft}
\alias{warcraft}
\title{Play a short sound from warcraft}
\usage{
warcraft(sound = 1, expr = NULL)
}
\arguments{
\item{sound}{character string or number specifying what sound to be played
by either specifying one of the built in sounds, specifying the path to a
wav file or specifying an url. The default is 1. Possible sounds are:
\enumerate{ \item \code{"jobsdone"} \item \code{"hello"} \item
\code{"peasant"} \item \code{"err"} \item \code{"peon"} \item
\code{"notready"} \item \code{"complete"} \item \code{"gold"} \item
\code{"blacksmith"} \item \code{"lumber"} }
If \code{sound} does not match any of the sounds above, or is a valid path
or url, a random sound will be played. Currently \code{warcraft} can only
handle http urls, https is not supported.}

\item{expr}{An optional expression to be excecuted before the sound.}
}
\description{
\code{warcraft} plays a short sound which is useful if you want to get
  notified, for example, when a script has finished. As an added bonus
  there are a number of different sounds to choose from.
}
\details{
If \code{warcraft} is not able to play the sound a warning is issued rather
  than an error. This is in order to not risk aborting or stopping the
  process that you wanted to get notified about.
}
\examples{
# Play an "jobsdone" sound
warcraft()

\dontrun{
# Play "hello" instead of a "jobsdone".
warcraft("hello")
# or
warcraft(2)

# Play a random sound
warcraft(0)

# Update all packages and "ping" when it's ready
update.packages(ask=FALSE); warcraft()
}
}
