% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analytics.R
\name{Analytics}
\alias{Analytics}
\title{Play a short sound from Analytics (group)}
\usage{
Analytics(sound = 1, expr = NULL)
}
\arguments{
\item{sound}{character string or number specifying what sound to be played by 
either specifying one of the built in sounds, specifying the path to a wav 
file or specifying an url. The default is 1. Possible sounds are:
\enumerate{ \item \code{"all_done"} }
If \code{sound} does not match any of the sounds above, or is a valid path
or url, a random sound will be played. Currently \code{Analytics} can only handle
http urls, https is not supported.}

\item{expr}{An optional expression to be excecuted before the sound.}
}
\description{
\code{Analytics} plays a short sound which is useful if you want to get notified, 
for example, when a script has finished. As an added bonus there are a number
of different sounds to choose from.
}
\details{
If \code{Analytics} is not able to play the sound a warning is issued rather than 
an error. This is in order to not risk aborting or stopping the process that
you wanted to get notified about.
}
\examples{
# Play an "all_done" sound
Analytics()

\dontrun{
# Play \\code{"all_done"} instead of a \\code{"fanfare"}.
Analytics("all_done")
# or
Analytics(1)

# Play a random sound
Analytics(0)

# Update all packages and "ping" when it's ready
update.packages(ask=FALSE); Analytics()
}
}
