% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Warcraft.R
\name{Warcraft}
\alias{Warcraft}
\title{Play a short sound from Warcraft}
\usage{
Warcraft(sound = 1, expr = NULL)
}
\arguments{
\item{sound}{character string or number specifying what sound to be played by 
either specifying one of the built in sounds, specifying the path to a wav 
file or specifying an url. The default is 1. Possible sounds are:
\enumerate{ \item \code{"jobsdone"} \item \code{"hello"} \item
\code{"peasant"} \item \code{"err"} \item \code{"peon"} \item
\code{"notready"} \item \code{"complete"} \item \code{"gold"} \item
\code{"blacksmith"} \item \code{"lumber"} }
If \code{sound} does not match any of the sounds above, or is a valid path
or url, a random sound will be played. Currently \code{Warcraft} can only handle
http urls, https is not supported.}

\item{expr}{An optional expression to be excecuted before the sound.}
}
\description{
\code{Warcraft} plays a short sound which is useful if you want to get notified, 
for example, when a script has finished. As an added bonus there are a number
of different sounds to choose from.
}
\details{
If \code{Warcraft} is not able to play the sound a warning is issued rather than 
an error. This is in order to not risk aborting or stopping the process that
you wanted to get notified about.
}
\examples{
# Play an "jobsdone" sound
Warcraft()

\dontrun{
# Play "hello" instead of a "jobsdone".
Warcraft("hello")
# or
Warcraft(2)

# Play a random sound
Warcraft(0)

# Update all packages and "ping" when it's ready
update.packages(ask=FALSE); Warcraft()
}
}
